Refactor svg-length and improve testing
We want to make sure we are getting the right values from our
svg-length parser and so we refactor the code to make it easer to
test and remove some of the cruft that had built up over the years.

Adjusted the lpe buildtool to detect the git repository.

+++ src/object/sp-offset.h
@@ -55,7 +55,7 @@ public:
 
     void *originalPath; ///< will be a livarot Path, just don't declare it here to please the gcc linker FIXME what?
     char *original;     ///< SVG description of the source path
-    float rad;          ///< offset radius
+    double rad;          ///< offset radius
 
     /// for interactive setting of the radius
     bool knotSet;
